Calculate CIA tod rate based on VIC settings
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@1061 75748be7-0387-4326-94c7-f41ce0883d2c

+++ sidplayfp/c64/CIA/mos6526.cpp
@@ -129,9 +129,9 @@ void MOS6526::clear(void)
 }
 
 
-void MOS6526::setDayOfTimeRate(double clock)
+void MOS6526::setDayOfTimeRate(unsigned int clock)
 {
-    m_todPeriod = (event_clock_t) (clock * (double) (1 << 7));
+    m_todPeriod = (event_clock_t) clock * (1 << 7);
 }
 
 void MOS6526::reset(void)
